requirements.txt
# requirements.txt
streamlit
google-generativeai
python-dotenv
bcrypt
pandas
matplotlib